pyproject.toml
[tool.poetry]
name = "wire"
version = "0.1.0"
description = "Your musicbot."
authors = ["illumfx <illumfx@mail.de>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.3"
python-dotenv = "^0.18.0"
pendulum = "^2.1.2"
humanize = "^3.9.0"
aiofiles = "^0.7.0"
asyncpg = "^0.23.0"
wavelink = "^0.9.9"
discord-ext-menus = {git = "https://github.com/Rapptz/discord-ext-menus.git"}
jishaku = "^2.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
